Rename MVVM folders to singular
