+++ Ballog/Model/TeamMember.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+struct TeamMember: Identifiable {
+    var id = UUID()
+    var name: String
+}
