+++ README.md
@@ -1,1 +1,4 @@
 # BallogApp
+
+## Build
+- `swiftc Ballog/View/*.swift Ballog/Model/*.swift Ballog/ViewModel/*.swift -o Ballog/build`
